1.0.17 新增F8Framework.F8ExcelDataClass和F8Framework.Launcher程序集，配置表生成的F8Framework.F8ExcelDataClass可用于HybridCLR热更，增加导表后脚本编译完成，再打包AB步骤，重新加载C#域后，AssetDatabase.Refresh需要增加delayCall延迟执行，完善热更文档，F8Framework.Tests程序集增加宏定义不参与打包

+++ Editor/ExcelTool/ExcelDataTool.cs
@@ -9,6 +9,8 @@ using System.Text;
 using UnityEngine;
 using UnityEditor;
 using Excel;
+using UnityEditor.Compilation;
+using Assembly = System.Reflection.Assembly;
 #if UNITY_WEBGL
 using LitJson;
 #else
@@ -29,13 +31,13 @@ namespace F8Framework.Core.Editor
         public const string ExcelPath = "/StreamingAssets/config"; //需要导表的目录
         public const string DLLFolder = "/F8Framework/ConfigData"; //存放dll目录
         public const string FileIndexFile = "config/fileindex.txt"; //fileindex文件目录
-        private static List<string> codeList; //存放所有生成的类的代码
+        private static Dictionary<string, string> codeList; //存放所有生成的类的代码
 
         private static Dictionary<string, List<ConfigData[]>> dataDict; //存放所有数据表内的数据，key：类名  value：数据
 
         // 使用StringBuilder来优化字符串的重复构造
         private static StringBuilder FileIndex = new StringBuilder();
-
+        
         private static string GetScriptPath()
         {
             MonoScript monoScript = MonoScript.FromScriptableObject(CreateInstance<ExcelDataTool>());
@@ -51,13 +53,26 @@ namespace F8Framework.Core.Editor
         
         private static void CreateAsmdefFile()
         {
-            // 创建.asmref文件的路径
-            string asmrefPath = Application.dataPath + DataManagerFolder + "/F8DataManager.asmref";
+            // 创建.asmdef文件的路径
+            string asmrefPath = Application.dataPath + DLLFolder + "/" + CODE_NAMESPACE + ".asmdef";
             
             FileTools.CheckFileAndCreateDirWhenNeeded(asmrefPath);
             // 创建一个新的.asmdef文件
             string asmdefContent = @"{
-    ""reference"": ""F8Framework.Core""
+    ""name"": ""F8Framework.F8ExcelDataClass"",
+    ""references"": [
+        ""F8Framework.Core"",
+        ""LitJson""
+    ],
+    ""includePlatforms"": [],
+    ""excludePlatforms"": [],
+    ""allowUnsafeCode"": false,
+    ""overrideReferences"": false,
+    ""precompiledReferences"": [],
+    ""autoReferenced"": false,
+    ""defineConstraints"": [],
+    ""versionDefines"": [],
+    ""noEngineReferences"": false
 }";
 
             // 将内容写入.asmdef文件
@@ -88,17 +103,19 @@ namespace F8Framework.Core.Editor
             }
             
             string F8DataManagerPath = FileTools.FormatToUnityPath(FileTools.TruncatePath(GetScriptPath(), 3)) + "/ConfigData/F8DataManager";
-            FileTools.SafeDeleteDir(F8DataManagerPath);
+            FileTools.SafeClearDir(F8DataManagerPath);
             FileTools.CheckDirAndCreateWhenNeeded(F8DataManagerPath);
             string F8ExcelDataClassPath = FileTools.FormatToUnityPath(FileTools.TruncatePath(GetScriptPath(), 3)) + "/ConfigData/F8ExcelDataClass";
-            FileTools.SafeDeleteDir(F8ExcelDataClassPath);
+            FileTools.SafeClearDir(F8ExcelDataClassPath);
             FileTools.CheckDirAndCreateWhenNeeded(F8ExcelDataClassPath);
+            string F8ExcelDataClassPathDLL = FileTools.FormatToUnityPath(FileTools.TruncatePath(GetScriptPath(), 3)) + "/ConfigData/" + CODE_NAMESPACE + ".asmdef";
+            FileTools.SafeDeleteFile(F8ExcelDataClassPathDLL);
             CreateAsmdefFile();
             AssetDatabase.Refresh();
             
             if (codeList == null)
             {
-                codeList = new List<string>();
+                codeList = new Dictionary<string, string>();
             }
             else
             {
@@ -113,8 +130,7 @@ namespace F8Framework.Core.Editor
             {
                 dataDict.Clear();
             }
-
-            float step = 1f;
+            
             FileIndex.Clear();
             FileTools.SafeDeleteFile(URLSetting.CS_STREAMINGASSETS_URL + FileIndexFile);
             FileTools.SafeDeleteFile(URLSetting.CS_STREAMINGASSETS_URL + FileIndexFile + ".meta");
@@ -122,7 +138,6 @@ namespace F8Framework.Core.Editor
             FileTools.CheckFileAndCreateDirWhenNeeded(URLSetting.CS_STREAMINGASSETS_URL + FileIndexFile);
             foreach (string item in files)
             {
-                step++;
                 GetExcelData(item);
                 OnLogCallBack(item.Substring(item.LastIndexOf('\\') + 1));
             }
@@ -132,29 +147,81 @@ namespace F8Framework.Core.Editor
                 EditorUtility.DisplayDialog("注意！！！", "\n暂无可以导入的数据表！", "确定");
                 throw new Exception("暂无可以导入的数据表！");
             }
-
             //编译代码,生成包含所有数据表内数据类型的dll
-            Assembly assembly = CompileCode(codeList.ToArray());
+            GenerateCodeFiles(codeList);
+            ScriptGenerator.CreateDataManager(codeList);
             AssetDatabase.Refresh();
+            // 等待脚本编译完成
+            EditorPrefs.SetBool("compilationFinished", false);
+            CompilationPipeline.compilationFinished += (object s) =>
+            {
+                EditorPrefs.SetBool("compilationFinished", true);
+            };
+        }
+
+        // 等待脚本编译完成
+        [UnityEditor.Callbacks.DidReloadScripts]
+        private static void AllScriptsReloaded()
+        {
+            if (EditorPrefs.GetBool("compilationFinished", false) == false)
+            {
+                return;
+            }
+            EditorPrefs.SetBool("compilationFinished", false);
+            LogF8.LogConfig("<color=#FF9E59>导表后脚本编译完成!</color>");
+            Assembly assembly = Util.Assembly.GetAssembly(CODE_NAMESPACE);
             //准备序列化数据
             string BinDataPath = Application.dataPath + BinDataFolder; //序列化后的数据存放路径
             if (Directory.Exists(BinDataPath)) Directory.Delete(BinDataPath, true); //删除旧的数据文件
             Directory.CreateDirectory(BinDataPath);
-            step = 1;
+            
+            string INPUT_PATH = Application.dataPath + ExcelPath;
+            var files = Directory.GetFiles(INPUT_PATH, "*.*", SearchOption.AllDirectories)
+                .Where(s => s.EndsWith(".xls") || s.EndsWith(".xlsx")).ToArray();
+            if (codeList == null)
+            {
+                codeList = new Dictionary<string, string>();
+            }
+            else
+            {
+                codeList.Clear();
+            }
+            if (dataDict == null)
+            {
+                dataDict = new Dictionary<string, List<ConfigData[]>>();
+            }
+            else
+            {
+                dataDict.Clear();
+            }
+            foreach (string item in files)
+            {
+                GetExcelData(item);
+            }
+            
             foreach (KeyValuePair<string, List<ConfigData[]>> each in dataDict)
             {
-                step++;
                 //Assembly.CreateInstance 方法 (String) 使用区分大小写的搜索，从此程序集中查找指定的类型，然后使用系统激活器创建它的实例化对象
                 object container = assembly.CreateInstance(CODE_NAMESPACE + "." + each.Key);
                 Type temp = assembly.GetType(CODE_NAMESPACE + "." + each.Key + "Item");
                 //序列化数据
                 Serialize(container, temp, each.Value, BinDataPath);
             }
-            ScriptGenerator.CreateDataManager(assembly);
-            AssetDatabase.Refresh();
+            
+            // 如果 Unity 检测到任何脚本更改，则会重新加载 C# 域。这样做的原因是可能已创建新的脚本化导入器 (Scripted Importer)，
+            // 它们的逻辑可能会影响“刷新”队列中的资源导入结果。此步骤会重新启动 Refresh() 以确保所有新的脚本化导入器生效。
+            UnityEditor.EditorApplication.delayCall += () =>
+            {
+                AssetDatabase.Refresh();
+                if (EditorPrefs.GetBool("compilationFinishedBuildAB", false) == true)
+                {
+                    ABBuildTool.BuildAllAB();
+                }
+                EditorPrefs.SetBool("compilationFinishedBuildAB", false);
+            };
             LogF8.LogConfig("<color=yellow>导表成功!</color>");
         }
-
+        
         [UnityEditor.MenuItem("开发工具/运行时读取Excel _F7", false, 101)]
         public static void ReLoadExcelData()
         {
@@ -267,8 +334,12 @@ namespace F8Framework.Core.Editor
                     {
                         //根据刚才的数据来生成C#脚本
                         ScriptGenerator generator = new ScriptGenerator(inputPath, className, names, types);
-                        //所有生成的类的代码最终保存在这个链表中
-                        codeList.Add(generator.Generate());
+                        //所有生成的类的代码最终保存在这
+                        if (codeList.ContainsKey(className))
+                        {
+                            throw new Exception("类名重复: " + className + " ,路径:  " + inputPath);
+                        }
+                        codeList.Add(className, generator.Generate());
                         if (dataDict.ContainsKey(className))
                         {
                             throw new Exception("类名重复: " + className + " ,路径:  " + inputPath);
@@ -313,6 +384,21 @@ namespace F8Framework.Core.Editor
             return cr.CompiledAssembly;
         }
 
+        // 生成代码文件
+        public static void GenerateCodeFiles(Dictionary<string, string> codeList)
+        {
+            string path = Application.dataPath + DLLFolder + "/F8ExcelDataClass";
+            FileTools.SafeClearDir(path);// 删除旧文件
+
+            // 将每个脚本写入独立的 .cs 文件
+            foreach (var kvp in codeList)
+            {
+                string filePath = $"{path}/{kvp.Key}.cs";
+                File.WriteAllText(filePath, kvp.Value);
+                LogF8.LogConfig($"已生成代码 " + path + "/<color=#FF9E59>" + kvp.Key + "</color>");
+            }
+        }
+        
         //序列化对象
         private static void Serialize(object container, Type temp, List<ConfigData[]> dataList, string BinDataPath)
         {
@@ -320,7 +406,7 @@ namespace F8Framework.Core.Editor
             foreach (ConfigData[] datas in dataList)
             {
                 //Type.FullName 获取该类型的完全限定名称，包括其命名空间，但不包括程序集。
-                object t = temp.Assembly.CreateInstance(temp.FullName);
+                object t = Util.Assembly.GetTypeInstance(temp.FullName);
                 foreach (ConfigData data in datas)
                 {
                     //Type.GetField(String) 搜索Type内指定名称的公共字段。
@@ -360,12 +446,23 @@ namespace F8Framework.Core.Editor
             // 记录日志
             LogF8.LogConfig("已序列化 " + BinDataPath + "/<color=#FFFF00>" + container.GetType().Name + ".json</color>");
 #else
-            IFormatter f = new BinaryFormatter();
-            Stream s = new FileStream(BinDataPath + "/" + container.GetType().Name + ".bytes", FileMode.OpenOrCreate,
-                FileAccess.Write, FileShare.Write);
-            f.Serialize(s, container);
-            LogF8.LogConfig("已序列化 " + BinDataPath + "/<color=#FFFF00>" + container.GetType().Name + ".bytes</color>");
-            s.Close();
+            try
+            {
+                IFormatter formatter = new BinaryFormatter();
+                string filePath = Path.Combine(BinDataPath, container.GetType().Name + ".bytes");
+
+                // 使用 using 语句确保流被正确关闭
+                using (Stream stream = new FileStream(filePath, FileMode.OpenOrCreate, FileAccess.Write, FileShare.None))
+                {
+                    formatter.Serialize(stream, container);
+                }
+
+                LogF8.LogConfig($"已序列化 {BinDataPath}/<color=#FFFF00>{container.GetType().Name}.bytes</color>");
+            }
+            catch (Exception ex)
+            {
+                Debug.LogError($"序列化失败: {ex.Message}");
+            }
 #endif
         }
     }
